[NAE-1679] User list frontend component
- corrected according to PR
