+++ src/main/java/com/netgrif/application/engine/workflow/service/DataService.java
@@ -696,7 +696,7 @@ public class DataService implements IDataService {
     @Override
     public UserFieldValue makeUserFieldValue(String id) {
         IUser user = userService.resolveById(id, true);
-        return new UserFieldValue(user.getStringId(), user.getName(), user.getSurname(), user.getEmail());
+        return new UserFieldValue(user);
     }
 
     private void updateDataset(Case useCase) {
